[ADDED] AssociationAccessors concern #3
This concern can be added to any model to automatically add getters and
setters for desired associations.

Simply include the concern and then a function association_params which
returns a hash of association names and attributes lists like so:

```ruby
{
  :association_one => [:attr1, :attr2, :attr3],
  :association_two => [:attr4, :attr5, :attr6],
}
```
Just note that if associations have attributes with the same name then
the last association’s attribute will overwrite any previous ones with
the same name.